Fix text wrapping in buttons in some translations
Fixes https://github.com/tootsuite/mastodon/issues/1686. 
The max-width is not 100% necessary, but makes the button in the translation shown in that issue look better (without the max-width I'm adding, it is too wide to fit in the compose page) and not cause horizontal scrolling issues on small screens. 
Does pose an issue if there are ever any buttons wider than 75px that don't have that max-width overridden with their own class or inline style - but I don't think there currently are any.